pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "hexlet-code"
authors = ["Mouzilo <6473839@gmail.com>"]
readme = "README.md"
packages = [{include = "gendiff"}]

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
argparse = "^1.4.0"
flake8 = "^6.0.0"
pytest = "^7.3.2"
pytest-cov = "^4.1.0"
pyyaml = "^6.0"


[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
